warning 해결

+++ src/choco_2/choco/parallel/schedule.cpp
@@ -73,7 +73,7 @@ namespace parallel{
 				}
 				
 				task.tick -=
-					GetTickCount64() - tick;
+					(int)(GetTickCount64() - tick);
 
 				if(erase_it)
 					it = tasks.erase(it);
